Install bower and grunt.

+++ manifests/site.pp
@@ -57,6 +57,7 @@ define download ($source, $path = $name, $timeout = 300) {
     command => "curl -s -o $path $source",
     creates => $path,
     timeout => $timeout,
+    unless  => ["test -f $path"]
   }
 }
 
@@ -81,6 +82,8 @@ node default {
   # node versions
   include nodejs::v0_10
   class { 'nodejs::global': }
+  nodejs::module {"bower":      node_version => 'v0.10' }
+  nodejs::module {"grunt-cli":  node_version => 'v0.10' }
 
   # default ruby versions
   include ruby::2_0_0_p353
